hopefully fix MSVC build for real

+++ src/v_trans.c
@@ -210,7 +210,7 @@ byte V_Colorize (byte *playpal, int cr, byte source, boolean keepgray109)
         }
         else if (cr == CR_BLUE || cr == CR_BLUE2)
         {
-            hsv.x = 240.f/360.;
+            hsv.x = (float) 240./360.;
 
             if (cr == CR_BLUE)
             {
@@ -220,7 +220,7 @@ byte V_Colorize (byte *playpal, int cr, byte source, boolean keepgray109)
         }
         else if (cr == CR_ORANGE)
         {
-            hsv.x = 26.f/360.;
+            hsv.x = (float) 26./360.;
         }
      }
